all new

+++ lib/Manager/manager_home.dart
@@ -7,8 +7,10 @@ import 'package:furniture_ecom_app/Manager/total_admin.dart';
 import 'package:furniture_ecom_app/Manager/total_marketers.dart';
 import 'package:furniture_ecom_app/Manager/dealers_list_manager.dart';
 import 'package:furniture_ecom_app/core/api/admin_api_service.dart';
+import 'package:furniture_ecom_app/core/services/user_service.dart';
 import 'package:furniture_ecom_app/my_ecom/animations/animation.dart';
 import 'package:furniture_ecom_app/my_ecom/my_constants/colors.dart';
+import 'package:furniture_ecom_app/my_ecom/my_constants/snackbar.dart';
 import 'package:furniture_ecom_app/my_login_screen.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -101,7 +103,6 @@ class _ManagerHomeState extends State<ManagerHome> {
             backgroundColor: Colors.transparent,
             elevation: 0,
             centerTitle: true,
-
           ),
         ),
       ),
@@ -300,8 +301,7 @@ class _ManagerHomeState extends State<ManagerHome> {
                                         Navigator.push(
                                           context,
                                           MaterialPageRoute(
-                                            builder: (_) =>
-                                                const TotalUsersM(),
+                                            builder: (_) => const TotalUsersM(),
                                           ),
                                         );
                                       },
@@ -385,14 +385,20 @@ class ManagerDrawer extends StatelessWidget {
   const ManagerDrawer({super.key, required this.currentPage});
 
   Future<void> _logout(BuildContext context) async {
-    final prefs = await SharedPreferences.getInstance();
-    await prefs.clear();
+    final result = await UserService.logout();
 
-    Navigator.pushAndRemoveUntil(
-      context,
-      MaterialPageRoute(builder: (_) => const MyLoginScreen()),
-      (route) => false,
-    );
+    if (result["success"] == true) {
+      final prefs = await SharedPreferences.getInstance();
+      await prefs.clear();
+
+      Navigator.pushAndRemoveUntil(
+        context,
+        MaterialPageRoute(builder: (_) => const MyLoginScreen()),
+        (route) => false,
+      );
+    } else {
+      showTopSnackBar(context, result["message"]);
+    }
   }
 
   @override
@@ -467,7 +473,6 @@ class ManagerDrawer extends StatelessWidget {
           ),
           const SizedBox(height: 10),
 
-          const SizedBox(height: 10),
           _drawerItem(
             context,
             Icons.people,
@@ -581,6 +586,7 @@ class ManagerDrawer extends StatelessWidget {
           title,
           style: GoogleFonts.poppins(
             color: isSelected ? mythemecolor : Colors.black87,
+            fontSize: isTablet(context) ? 14 : 12,
             fontWeight: isSelected ? FontWeight.w600 : FontWeight.normal,
           ),
         ),
